Rename `dxw_members_only_redirect` action to `dxw_members_only`
Before, we had both an action and a filter named
`dxw_members_only_redirect`. Under the hood, WordPress filters and
actions are actually the same thing (`add_action` is actually just a
wrapper for `add_filter`), so you can't have an action and a filter that
share the same name but are intended to behave independently.

Therefore, I've renamed the `dxw_members_only_redirect` action (which
actually just indicates that the plugin has started to operate) to just
`dxw_members_only`.

I searched our codebase to check that this hook was not in use anywhere,
and couldn't find any references. So though this is a breaking change,
it should not break anything on any of our own sites.