Add test for redirect bug that requires fixing
There is currently a bug in the plugin where, if you visit a URL that
requires a login whilst not logged in, the redirect to the login page is
served with a cache-control max-age of the default max age setting.

This can result in the redirect being cached in a user's browser, so
even if they successfully authenticate, when they attempt to visit the
URL again, the cached redirect is served to them, rather than the page
itself. Therefore, they end up in a loop of being repeatedly prompted to
log in.

The redirect should have max-age of 0. This change adds the test for the
expected functionality, and uses Playwright's `test.fail()` annotation
to indicate that the bug exists and requires fixing (as the test is
currently expected to fail).