Add test for buggy cache-control max-age once logged in
The intention of the plugin is that, once a user has successfully logged
in, they are served with a cache-control header of `private, max-age=x`
where `x` is the default max age set in the dxw Members Only settings
page.

However, WordPress sets the cache-control header for authenticated users
later than the plugin hooks in, so this gets over-ridden with the
default WordPress cache-control header for all logged in users of
`private, max-age=0`.

This failing test documents that bug.