Ensure we wait for login process to complete before making assertions
In a couple of tests, we were submitting the login form then relying on
the standard redirect behaviour to take us to the required destination
page and making assertions about that destination page.

We'd get intermittent failures because Playwright would start making
those assertions before the login and redirect process and completed.
Now, we wait until we're at the destination URL before making the
assertions.