Do `npx playwright install --with-deps` as part of bootstrapping
I'd assumed that in a non-CI environment, just `npm install` would be
enough to get all the packages that Playwright needs (including e.g.
browser binaries). But it turns out that's not the case, we also need to
run `npx playwright install --with-deps` as well.

Because we're now running that in `script/bootstrap`, I've removed the
separate call that happens foir the same command in `script/cibuild`