Add e-2-e tests for upgrade migration process
This includes a test that documents an existing bug, where we set
default values for some options before we attempt to import the options
from their old `new-members-only` names to the new `dxw-members-only`
equivalent. In those cases, the import process views the options as
already populated, and therefore does not import the old values.