Move `upload.php` logic to `Upload` class and add unit tests
Note: both before and after this refactor, there are significant issues
with this functionality. Specifically:

- The checkbox to add the uploaded file to the allow list only appears
  when adding media directly to a post via the classic editor. It does
not appear in the block editor or when editing the file in the media
library
- Checking the box does not actually update the
  `dxw_members_only_list_content` option in order to make the file
publicly accessible (i.e. it makes not difference if the box is checked
or not, even when it does appear).